requirements.txt
ultralytics==8.0.114
pandas==2.0.2
opencv-python==4.5.4.60
numpy==1.24.3
scipy==1.10.1
easyocr==1.7.0
filterpy==1.4.5
lap==0.4.0
alembic==1.11.1
antiorm==1.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
db==0.1.1
db-sqlite3==0.0.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
pillow==9.5.0
